+++ src/ocean/util/log/Config.d
@@ -84,7 +84,7 @@ import ocean.transition;
 import ocean.io.Stdout;
 import ocean.core.Array : removePrefix, removeSuffix;
 import ocean.core.Array_tango : sort;
-import ocean.util.config.ClassFiller;
+import ocean.util.config.ConfigFiller;
 import ocean.util.config.ConfigParser;
 import ocean.util.log.AppendSysLog;
 import ocean.text.util.StringSearch;