+++ RELEASE_NOTES.md
@@ -8,3 +8,16 @@ tango      | v1.3.x
 
 New Features
 ============
+
+* `ocean.util.config.ConfigFiller`
+
+  Provides the same functionality as the old `ClassFiller`, but it's
+  extended to support `struct`s too.
+
+Deprecations
+============
+
+* `ocean.util.config.ClassFiller`
+
+  Deprecated in favour of the new `ConfigFiller` which provides the
+  same interface.